Part four exercises finished

+++ controllers/blogs.js
@@ -1,9 +1,13 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (_request, response) => {
-  const blogs = await Blog.find({})
-  response.status(200).json(blogs)
+  const blogs = await Blog.find({}).populate('user', {
+    username: 1,
+    name: 1,
+  })
+  response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -18,22 +22,46 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const { title, author, url, likes } = request.body
 
+  const decodedToken = request.user
+
+  const findOrCreateUser = async (userId) => {
+    if (userId) return await User.findById(userId)
+    return await User.findOne()
+  }
+
+  const user = await findOrCreateUser(decodedToken.id)
+
   const blog = new Blog({
     title,
     author,
     url,
     likes,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
+  const decodedToken = request.user
 
-  if (blog) return response.status(204).end()
-  response.status(404).end()
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({
+      error: 'you can only delete your own blogs',
+    })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
